Test Drive 2

+++ firmware/ota.py
@@ -24,7 +24,7 @@ class OTAUpdater:
         # get the current version (stored in version.json)
         if "version.json" in os.listdir():
             with open("version.json") as f:
-                self.current_version = int(json.load(f)["version"])
+                self.current_version = float(json.load(f)["version"])
             print(f"Current device firmware version is '{self.current_version}'")
 
         else:
@@ -93,7 +93,7 @@ class OTAUpdater:
         print(f"data is: {data}, url is: {self.version_url}")
         print(f"data version is: {data['version']}")
 
-        self.latest_version = int(data["version"])
+        self.latest_version = float(data["version"])
         print(f"latest version is: {self.latest_version}")
 
         # compare versions
